+++ tests/pyccel/scripts/import_syntax/from_mod_import_as_user_func.py
@@ -1,6 +1,6 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 
-def fun(xi1 : 'double', xi2 : 'double', xi3 : 'double'):
+def fun(xi1 : 'float', xi2 : 'float', xi3 : 'float'):
     from user_mod import user_func as f
     return f(xi1, xi2, xi3)
