+++ tests/pyccel/scripts/import_syntax/import_mod_as.py
@@ -1,7 +1,7 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 import numpy as np
 
-#$ header function matmat(double [:,:], double [:,:], double [:,:])
+#$ header function matmat(float [:,:], float [:,:], float [:,:])
 def matmat(a,b,c):
     n, m = np.shape(a)
     m, p = np.shape(b)
@@ -16,8 +16,8 @@ if __name__ == '__main__':
     m = 4
     p = 3
 
-    a = np.zeros((n,m), 'double')
-    b = np.zeros((m,p), 'double')
+    a = np.zeros((n,m), 'float')
+    b = np.zeros((m,p), 'float')
 
     for i in range(0, n):
         for j in range(0, m):
@@ -33,7 +33,7 @@ if __name__ == '__main__':
         print()
     print()
 
-    c = np.zeros((n,p),'double')
+    c = np.zeros((n,p),'float')
     matmat(a,b,c)
 
     for i in range(0, n):
