+++ tests/pyccel/scripts/import_syntax/import_mod_user.py
@@ -1,7 +1,7 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 import user_mod
 
-def fun(xi1 : 'double', xi2 : 'double', xi3 : 'double'):
+def fun(xi1 : 'float', xi2 : 'float', xi3 : 'float'):
     return user_mod.user_func(xi1, xi2, xi3)
 
 if __name__ == '__main__':
