+++ tests/pyccel/scripts/import_syntax/user_mod.py
@@ -1,4 +1,4 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 
-def user_func(x1 : 'double', x2 : 'double', x3 : 'double'):
+def user_func(x1 : 'float', x2 : 'float', x3 : 'float'):
     return x1+x2+x3
